importFileParser function created. Task 5.3 done.
