+++ ImportService/handler.js
@@ -1,31 +1,4 @@
-import { S3 } from "aws-sdk"
+import { importProductsFile } from "./importProductsFile/importProductsFile";
+import { importFileParser } from "./importFileParser/importFileParser";
 
-const response = (statusCode, payload) => {
-    return {
-        statusCode,
-        headers: {
-            'Access-Control-Allow-Origin': '*',
-            'Content-Type': 'application/json',
-        },
-        body: payload
-    }
-}
-
-const s3 = new S3({ region: 'eu-north-1' })
-const BUCKET = 'import-service-bucket-zafar'
-
-export const importProductFiles = async (event) => {
-
-    const name = event.queryStringParameters.name
-
-    const params = {
-        Bucket: BUCKET,
-        Key: `uploaded/${name}`,
-        Expires: 60,
-        ContentType: 'text/csv'
-    }
-
-    const signedUrl = await s3.getSignedUrlPromise('putObject', params)
-
-    return response(200, signedUrl)
-}
+export { importProductsFile, importFileParser }
